GHC.HaskellToCore.Bindings

Description

Pattern-matching bindings (HsBinds and MonoBinds)

Handles GHC.Haskell.Bindings; those at the top level require different handling, in that the RecNonRecetc. structure is thrown away (whereas at lower levels it is preserved with let/letrecs).

NB: Supplants DsBinds.

Documentation

module DsBinds