GHC.Plugins

Description

This module is not used by GHC itself. Rather, it exports all of the functions and types you are likely to need when writing a plugin for GHC. So authors of plugins can probably get away simply with saying import GHC.GHC.Plugins.

Particularly interesting modules for plugin writers include GHC.Core and GHC.Core.Optimize.Monad.

NB: Supplants GhcPlugins.

Documentation

getTyVar :: Type -> Maybe TyVar

NB: Renamed from getTyVar_maybe.

nameModule :: Name -> Maybe Module

NB: Renamed from nameModule_maybe.

splitAppTy :: Type -> Maybe (Type, Type)

NB: Renamed from splitAppTy_maybe.

splitForAllTy :: Type -> Maybe (TyVar, Type)

NB: Renamed from splitForAllTy_maybe.

tyConAppTyCon :: Type -> Maybe TyCon

NB: Renamed from tyConAppTyCon_maybe.

tyConAppArgs :: Type -> Maybe [Type]

NB: Renamed from tyConAppArgs_maybe.