GHC.Typecheck.Types

Description

Various types used during typechecking, please see GHC.Typecheck.Renaming.Monad as well for operations on these types. You probably want to import it, instead of this module.

All the monads exported here are built on top of the same IOEnv monad. The monad functions like a Reader monad in the way it passes the environment around. This is done to allow the environment to be manipulated in a stack like fashion when entering expressions... etc.

For state that is global and should be returned at the end (e.g not part of the stack mechanism), you should use a TcRef (= IORef) to store them.

NB: Supplants TcRnTypes.

Documentation

module TcRnTypes