GHC.Types.Unique

Description

Uniques are used to distinguish entities in the compiler (Ids, Classes, etc.) from each other. Thus, Uniques are the basic comparison key in the compiler.

If there is any single operation that needs to be fast, it is Unique comparison. Unsurprisingly, there is quite a bit of huff-and-puff directed to that end.

Some of the other hair in this code is to be able to use a splittable UniqSupply if requested/possible (not standard Haskell).

NB: Supplants Unique.

Documentation

module Unique